backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
pdfplumber==0.10.3
python-dotenv==1.0.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
